+++ CHANGES.md
@@ -1,5 +1,12 @@
 # Changes
 
+# v0.2.0
+
+- Add onDispatch event to module definition
+- Add onDispatch function to log helpers
+- Use onDispatch in testForm example
+- Fix bug when call dispatch from a child component
+
 # v0.1.0
 
 - Add mapToObj helper